requirements.txt
python-telegram-bot==20.2
aioredis==2.0.1
hiredis==2.0.0
marshmallow==3.18.0
aiohttp-apispec==2.2.3
aiohttp==3.8.3